+++ server/src/Birdism/Api.hs
@@ -3,7 +3,6 @@ module Birdism.Api where
 import           Control.Lens
 import           Servant
 
-
 import qualified Data.Aeson        as J
 import qualified Data.Aeson.Casing as J
 import qualified Data.Text         as T
@@ -13,6 +12,8 @@ import           Birdism.Config
 import           Birdism.Lib
 import           Birdism.Types
 import           GHC.TypeLits      (Nat)
+import qualified Birdism.Cache as Cache
+import Birdism.Cache (HasBirdismCache(birdismCache))
 
 
 type BirdismHttpAPIF f
@@ -90,11 +91,11 @@ processSearch
   :: ( MonadIO m
      , MonadReader r m
      , HasDbConfig r
-     , HasFlickrContext r
+     , HasBirdismCache r
+     , HasFlickrConf r
      , HasEBirdConf r
      , MonadError e m
      , AsEbirdError e
-     , HasAppCtx r
      )
   => SearchRequest -> m SearchResult
 processSearch (SearchRequest regionCode familySciName) = do
@@ -107,7 +108,7 @@ processSpeciesSearch
      , MonadReader r m
      , HasDbConfig r
      , HasEBirdConf r
-     , HasAppCtx r
+     , HasBirdismCache r
      , MonadError e m
      , AsEbirdError e
      )
@@ -119,7 +120,8 @@ processSpeciesSearch (SearchRequest regionCode familySciName) = do
 
 processImageSearch
   :: ( MonadReader r m
-     , HasFlickrContext r
+     , HasFlickrConf r
+     , HasBirdismCache r
      , MonadIO m
      )
   => [CommonName] -> m SearchResult
@@ -145,30 +147,33 @@ instance J.ToJSON FamilyScientificNameResponse where
 getFamilyScientificName
   :: ( MonadIO m
      , MonadReader r m
-     , HasAppCtx r
+     , HasBirdismCache r
      )
   => FamilyScientificNameRequest -> m FamilyScientificNameResponse
 getFamilyScientificName (FamilyScientificNameRequest commonName) = do
-  fams <- asks (^. axBirdFamiliesCache)
+  cache <- asks (^. birdismCache)
+  families <- Cache.getFamiliesCache cache
   let isSubStr v1 v2 = T.isInfixOf (T.toLower $ uCommonName v1) (T.toLower $ uCommonName v2)
-      found = filter (isSubStr commonName . _fCommonName) (unFamiliesCache fams)
+      found = filter (isSubStr commonName . _fCommonName) families
   pure $ FamilyScientificNameResponse found
 
 validateRegion
-  :: (MonadReader s m, HasAppCtx s, MonadError e m, AsEbirdError e)
+  :: (MonadReader r m, HasBirdismCache r, MonadError e m, AsEbirdError e, MonadIO m)
   => RegionCode -> m Region
 validateRegion region = do
-  cache <- unRegionsCache <$> asks (^. axRegionsCache)
+  cache <- asks (^. birdismCache)
+  regions <- Cache.getRegionsCache cache
   let findRegion x = region == _rRegionCode x
-  validate findRegion cache (_EbirdErrorSearch # "Invalid region")
+  validate findRegion regions (_EbirdErrorSearch # "Invalid region")
 
 validateFamily
-  :: (MonadReader s m, HasAppCtx s, MonadError e m, AsEbirdError e)
+  :: (MonadIO m, MonadReader r m, HasBirdismCache r, MonadError e m, AsEbirdError e)
   => ScientificName -> m Family
 validateFamily family = do
-  cache <- unFamiliesCache <$> asks (^. axBirdFamiliesCache)
+  cache <- asks (^. birdismCache)
+  families <- Cache.getFamiliesCache cache
   let findScName x = family == _fScientificName x
-  validate findScName cache (_EbirdErrorSearch # "Invalid family")
+  validate findScName families (_EbirdErrorSearch # "Invalid family")
 
 -- | Function that generalizes some validation and parsing of string-ly values
 -- into proper domain-types. E.g. - a string based region code can be converted
@@ -179,21 +184,22 @@ validate check cache e = maybe (throwError e) pure $ find check cache
 getFamilies
   :: ( MonadIO m
      , MonadReader r m
-     , HasAppCtx r
+     , HasBirdismCache r
      )
   => m FamilyNames
-getFamilies =
-  asks (^. axBirdFamiliesCache) <&> unFamiliesCache
+getFamilies = do
+  cache <- asks (^. birdismCache)
+  Cache.getFamiliesCache cache
 
 getRegions
   :: ( MonadIO m
      , MonadReader r m
-     , HasAppCtx r
+     , HasBirdismCache r
      )
   => m RegionNames
-getRegions =
-  asks (^. axRegionsCache) <&> unRegionsCache
-
+getRegions = do
+  cache <- asks (^. birdismCache)
+  Cache.getRegionsCache cache
 
 -- * Internal Servant extensions