+++ server/src/Birdism/Init.hs
@@ -5,6 +5,7 @@ module Birdism.Init where
 import qualified Data.ByteString.Char8      as BC
 import qualified Data.Text.Encoding         as T
 import qualified Database.PostgreSQL.Simple as PG
+import qualified Birdism.Cache as Cache
 
 import           Data.FileEmbed             (embedFile)
 import           Data.String                (fromString)
@@ -12,7 +13,6 @@ import           Data.String                (fromString)
 import           Birdism.Common
 import           Birdism.Config
 import           Birdism.Data
-import           Service.Flickr.Context     (FlickrContext (..), mkFlickrCache)
 
 initialiseAppCtx :: MonadIO m => AppConfig -> m AppCtx
 initialiseAppCtx (AppConfig dbUrl port ebird flickr) = do
@@ -20,11 +20,10 @@ initialiseAppCtx (AppConfig dbUrl port ebird flickr) = do
   let dbConf = DbConfig conn 10 1
   initialiseDatabase conn
   -- create an in-memory cache of family names and region, as they don't change
-  families <- runReaderT getFamilyNames dbConf
   regions  <- runReaderT getRegionNames dbConf
-  flickrCache <- mkFlickrCache
-  let flickrCtx = FlickrContext flickr flickrCache
-  return $ AppCtx dbConf (fromMaybe defaultServerPort port) (FamiliesCache families) (RegionsCache regions) ebird flickrCtx
+  families <- runReaderT getFamilyNames dbConf
+  birdismCache <- Cache.newCache regions families
+  return $ AppCtx dbConf (fromMaybe defaultServerPort port) birdismCache ebird flickr
 
 initialiseDatabase :: MonadIO m => PG.Connection -> m ()
 initialiseDatabase conn = do